backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
typing-extensions==4.12.2
pyyaml==6.0.2
redis==5.0.8
celery==5.4.0
chromadb==0.6.3
langgraph==0.2.35
python-jose==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
python-multipart==0.0.9
pytest==8.3.5
pytest-asyncio==0.23.8
voyageai==0.3.2

